introduce flowercluster/init
it's important for us to have the configuration for how the containers
were built or started on the machine running them, and it doesn't make
much sense for us to build all our automatoin managing containers
*outside* of the instance they're running on.

flowercluster/init is a docker image that runs a custom python script to
build containers with a RoleID, and to start containers with a SecretID
wrap token. this is accomplished by communicating directly with the
docker daemon and with vault, using the docker socket and the requests
library, respectively.

config.yml aims to be the site configuration file that describes the
containers to be running. it will mostly replace the config.yml in the
fabric deploy scripts. it still makes sense to run fabric for some
administrative tasks, like unsealing the vault and populating secrets.

this script runs after vault has been unsealed, but before any other
containers are started.

goals for this feature branch include:
- updating vault policies
- fetching RoleIDs and SecretID tokens from vault
- building all containers to run on the host
- starting containers with a valid SecretID
- revoking the build token when finished